requirements.txt
streamlit==1.54.0
pandas==2.2.3
psycopg2-binary==2.9.9
SQLAlchemy==2.0.34
alembic==1.13.2
